GG-7677 Remove v1 Retrievals
