+++ README.md
@@ -5,6 +5,11 @@ auth-client
 Library for supporting user authorisation on microservices.
 
 ## Change History
+
+### v9.0.0 Feb 2026
+
+Remove v1 Retrievals, long since deprecated after 2.11.0-play-26, to assist One Login integration.
+
 ### v8.8.0 Nov 2025
 
 Removed unused trustedHelper retrieval.
@@ -373,6 +378,12 @@ or provide a mocking session cookie as demonstrated below
 ```
 
 For more testing examples head to `uk.gov.hmrc.auth.core.AuthConnectorSpec` or [Auth-Demo-Frontend]("https://github.com/hmrc/auth-demo-frontend")
+
+## Running the tests
+```
+sm2 --start AUTH_CLIENT_ALL
+sbt clean compile test it/test
+```
 
 
 ## License
